refactor: InputValidator 클래스 util 패키지 하위로 이동
