+++ src/main/java/lotto/util/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import java.util.Objects;
